refactor: move feature collection struct (and related structs) to own file
